pyproject.toml
[project]
name = "kvasir-capsule-loader"
version = "0.1.0"
description = " Inofficial dataset abstraction KvasirCapsule dataset"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "black>=25.12.0",
    "click>=8.3.1",
    "mypy>=1.19.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pytest>=9.0.2",
    "requests>=2.32.5",
    "ruff>=0.14.8",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
]
